+++ client/webserver/site/src/js/notifications.js
@@ -1,0 +1,23 @@
+export const IGNORE = 0
+export const DATA = 1
+export const POKE = 2
+export const SUCCESS = 3
+export const WARNING = 4
+export const ERROR = 5
+
+/*
+ * make constructs a new notification. The notification structure is a mirror of
+ * the structure of notifications sent from the web server.
+ * NOTE: I'm hoping to make this function obsolete, since errors generated in
+ * javascript should usually be displayed/cached somewhere better. For example,
+ * if the error is generated during submission of a form, the error should be
+ * displayed on or near the form itself, not in the notifications.
+ */
+export function make (subject, details, severity) {
+  return {
+    subject: subject,
+    details: details,
+    severity: severity,
+    stamp: new Date().getTime()
+  }
+}
